Solved problem 10
Searched the internet for an algorithm to find prime numbers and came
across multiple including Eratosthenes and Atkins which I tried
implementing. I couldn't get an Eratosthenes solution to work so went
with Atkins creating the algorithm based on the pseudocode on Wikipedia.

Quick and easy to find primes up to a limit and would have been great
for some of the past problems!